Filled in the gaps
Flushed out a lot of the scaffolding code that was created in the
first commit. Added a number of accessor functions to the various
battery properties, including getStatus, getChargeAsperc, and
isInitialized. Filled in the draw function which calls out to the
various other draw functions, drawText being the only one that has
been filled in. Added a fit function as well.