drawBattery Refactor
Both cases where a battery needs to be draw should use the same
code. Refactored the into drawEmptyBattery. A second call to either
fillin the battery based on charge level or put an X through it if it
is missing is done after. The battery also properly scales which it
didn't really do in the prior math.